Created a package and moved activities into that.
