+++ app/src/main/AndroidManifest.xml
@@ -8,16 +8,16 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".Activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
 
-        <activity android:name=".DashBoardActivity">
+        <activity android:name=".Activity.DashBoardActivity">
             <intent-filter>
-                <action android:name="zz.demoapp.DASHBOARDACTIVITY" />
+                <action android:name="zz.demoapp.activity.DASHBOARDACTIVITY" />
                 <category android:name="android.intent.category.DEBUG" />
             </intent-filter>
         </activity>
