+++ app/src/main/java/zz/demoapp/Activity/MainActivity.java
@@ -1,9 +1,11 @@
-package zz.demoapp;
+package zz.demoapp.Activity;
 
 import android.content.Intent;
 import android.os.Handler;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import zz.demoapp.R;
 
 public class MainActivity extends AppCompatActivity {
