+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="zz.demoapp.MainActivity">
+    tools:context="zz.demoapp.Activity.MainActivity">
 
     <TextView
         android:text="FlashScreen"
